requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
google-auth==2.39.0
idna==3.10
inflection==0.5.1
packaging==24.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.9.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
sqlparse==0.5.3
stripe==12.1.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
